#1441 SingleDatePropertySettingDialog: a new dialog for setting a date value.
